Do not overwrite exitCh if it is nil
In some cases, we return nil for the exitCh. This happens when there is
no data from Consul or when an error occurs. In those cases, we were
overwriting the parent's exitCh, effectively orphaning the channel. In
this, the subprocess would exit, but its exit was never propagated to
the parent.

Fixes GH-85