fix issue with secret/prefix entries with no path
Issue came up when user had a secret config section with no path set
(but had noprefix) and where setting the path on the cli. The
documentation on how this was handled was unclear so the user thought it
was a bug.

I fixed the issue so that their test would have worked by filtering out
the secret config entry with no path (as the path is required).

I also updated the documentation to help clarify how prefix and secret
(they use the same code) work.