have standard 'log' package output to hclog
consult-template is still using the standard log package

These tweaks limits it to logging just the message and handing that to
hclog. It makes it look mostly like it is an normal log entry. Only diff
is that it uses different notation for subsections.

Might be useful just to keep as it makes using the standard log package
work easily with hclog.