only setpgid on child calls if subshell is used
Makes it so that only commands that use 'sh -c' to wrap a shell-script
like command get setpgid set to ensure signal propagation to the process
being run in the subshell. Uses new consul-template/child exported
features to detect and set it (thus the CT version bump).

This is needed to fix issues caused by setpgid, primarily when the
sub-shell isn't used it still effects things like you can't run an
interactive bash setting as the command and possibly other shell
scripts.